extract PostgresClient from addProduct
